+++ src/agdb/collections/dictionary/dictionary_impl.rs
@@ -15,6 +15,7 @@ where
     pub(crate) phantom_data: PhantomData<T>,
 }
 
+#[allow(dead_code)]
 impl<T, Data> DictionaryImpl<T, Data>
 where
     T: Clone + Default + Eq + PartialEq + StableHash + Serialize,
