+++ src/agdb/collections/map/map_impl.rs
@@ -19,6 +19,7 @@ where
     pub(crate) phantom_data: PhantomData<(K, T)>,
 }
 
+#[allow(dead_code)]
 impl<K, T, Data> MapImpl<K, T, Data>
 where
     K: Clone + Default + Eq + Hash + PartialEq + StableHash + Serialize,
