+++ src/agdb/collections/map_common.rs
@@ -26,6 +26,7 @@ where
     pub(crate) phantom_data: PhantomData<(K, T)>,
 }
 
+#[allow(dead_code)]
 impl<K, T, Data> MapCommon<K, T, Data>
 where
     K: Clone + Default + Eq + Hash + PartialEq + StableHash + Serialize,
