+++ src/agdb/collections/multi_map/multi_map_impl.rs
@@ -16,6 +16,7 @@ where
     pub(crate) map_common: MapCommon<K, T, Data>,
 }
 
+#[allow(dead_code)]
 impl<K, T, Data> MultiMapImpl<K, T, Data>
 where
     K: Clone + Default + Eq + Hash + PartialEq + StableHash + Serialize,
