+++ src/agdb/collections/storage_dictionary.rs
@@ -18,6 +18,7 @@ use std::rc::Rc;
 pub type StorageDictionary<T, Data = StorageFile> =
     DictionaryImpl<T, DictionaryDataStorage<T, Data>>;
 
+#[allow(dead_code)]
 impl<T, Data> StorageDictionary<T, Data>
 where
     T: Clone + Default + Eq + PartialEq + StableHash + Serialize,
