+++ src/agdb/commands/cmd.rs
@@ -1,0 +1,7 @@
+use super::Commands;
+
+#[allow(dead_code)]
+pub(crate) struct Cmd {
+    pub(crate) commands: Commands,
+    pub(crate) executed: bool,
+}
