+++ src/agdb/graph_search/search_control.rs
@@ -1,3 +1,4 @@
+#[allow(dead_code)]
 pub enum SearchControl {
     Continue(bool),
     Finish(bool),
